+++ src/about.tcl
@@ -139,7 +139,9 @@ proc about::About {} {
     \u2022 <link-awthemes>awthemes $vawthemes</link-awthemes>\n\n \
     \u2022 <link-ale_themes>ale_themes</link-ale_themes>\n\n \
     \u2022 <link-tkcc>tkcc</link-tkcc>\n\n \
-    \u2022 <link-repl>tcl-repl</link-repl>\n"
+    \u2022 <link-repl>tcl-repl</link-repl>\n \
+    \n menus/.mnu v$al(MNUversion) \
+    \n alited.ini v$al(INIversion)"
   set tab2 [list General Packages "{fra - - 1 99 {-st nsew -rw 1 -cw 1}} {.TexPack - - - - {pack -side left -expand 1 -fill both} {-w $wmax -h 31 -rotext ::alited::AboutPack -tags ::alited::about::textTags}}" Acknowledgements "{fra - - 1 99 {-st nsew -rw 1 -cw 1}} {.TexAckn - - - - {pack -side left -expand 1 -fill both} {-w $wmax -h 34 -rotext ::alited::AboutAckn -tags ::alited::about::textTags}} {.sbv .texAckn L - - {pack -side right}}"]
   lassign [::baltip cget -shiftX] -> shiftX
   ::baltip configure -shiftX 10