+++ src/favor_ls.tcl
@@ -21,12 +21,14 @@ namespace eval ::alited::favor_ls {
 
 # ________________________ Form's buttons _________________________ #
 
-proc favor_ls::Save_favlist {} {
+proc favor_ls::Save_favlist {{saveini yes}} {
   # Saves favorites' list.
+  #   saveini - save templates in ini-file
 
   variable favlist
   variable favlistsaved
   set favlistsaved $favlist
+  if {$saveini} alited::ini::SaveIniPrj
 }
 #_______________________
 
@@ -209,6 +211,7 @@ proc favor_ls::Add {} {
     set msg [string map [list %n [llength $favlist]] $al(MC,favnew)]
     alited::Message2 $msg 3
   }
+  Save_favlist
   Focus $isel
 }
 #_______________________
@@ -233,6 +236,7 @@ proc favor_ls::Change {} {
     set favcont [lreplace $favcont $isel $isel $currents]
     set msg [string map [list %n [incr isel]] $al(MC,favupd)]
     alited::Message2 $msg 3
+    Save_favlist
   }
 }
 #_______________________
@@ -258,6 +262,7 @@ proc favor_ls::Delete {} {
   if {$llen>=0} {Select $isel}
   set msg [string map [list %n $nsel] $al(MC,favrem)]
   alited::Message2 $msg 3
+  Save_favlist
 }
 
 # ________________________ Ini data _________________________ #
@@ -301,7 +306,7 @@ proc favor_ls::GetIni {lines} {
     lappend favpla  [lindex $lines 1]
     lappend favcont [join $cont $::alited::EOL]
   }
-  Save_favlist
+  Save_favlist no
 }
 #_______________________
 
@@ -339,11 +344,11 @@ proc favor_ls::_create {} {
   $obDl2 paveWindow $win {
     {fraLbxFav - - 1 2 {-st nswe -pady 4} {}}
     {.fra - - - - {pack -side right -fill both} {}}
-    {.fra.buTAd - - - - {pack -side top -anchor n} {-takefocus 0 -com ::alited::favor_ls::Add -tip "Add a list of favorites" -image alimg_add-big -relief flat -highlightthickness 0}}
-    {.fra.buTChg - - - - {pack -side top} {-takefocus 0 -com ::alited::favor_ls::Change -tip "Change a list of favorites" -image alimg_change-big -relief flat -highlightthickness 0}}
-    {.fra.buTDel - - - - {pack -side top} {-takefocus 0 -com ::alited::favor_ls::Delete -tip "Delete a list of favorites" -image alimg_delete-big -relief flat -highlightthickness 0}}
+    {.fra.buTAd - - - - {pack -side top -anchor n} {-takefocus 0 -com ::alited::favor_ls::Add -tip "Add a list of favorites" -image alimg_add-big -relief flat -overrelief raised -highlightthickness 0}}
+    {.fra.buTChg - - - - {pack -side top} {-takefocus 0 -com ::alited::favor_ls::Change -tip "Change a list of favorites" -image alimg_change-big -relief flat -overrelief raised -highlightthickness 0}}
+    {.fra.buTDel - - - - {pack -side top} {-takefocus 0 -com ::alited::favor_ls::Delete -tip "Delete a list of favorites" -image alimg_delete-big -relief flat -overrelief raised -highlightthickness 0}}
     {.fra.v_ - - - - {pack -side top -expand 1 -fill y}}
-    {.fra.buTCur - - - - {pack -side top} {-takefocus 0 -com ::alited::favor_ls::GetCurrentList -tip "$alited::al(MC,currfavs)" -image alimg_heart-big -relief flat -highlightthickness 0}}
+    {.fra.buTCur - - - - {pack -side top} {-takefocus 0 -com ::alited::favor_ls::GetCurrentList -tip "$alited::al(MC,currfavs)" -image alimg_heart-big -relief flat -overrelief raised -highlightthickness 0}}
     {.LbxFav - - - - {pack -side left -expand 1 -fill both} {-h 10 -w 40 -lvar ::alited::favor_ls::favlist}}
     {.sbvFavs fraLbxFav.LbxFav L - - {pack -side left -fill y} {}}
     {fra1 fraLbxFav T 1 2 {-st nswe}}