+++ src/menu.tcl
@@ -250,7 +250,10 @@ proc menu::RunTip {} {
     ### ________________________ Other tools _________________________ ###
   $m add separator
   $m add command -label $al(MC,checktcl) -command alited::CheckRun
-  $m add command -label $al(MC,filelist) -command {alited::bar::BAR popList} -accelerator $al(acc_21)
+  menu $m.filelist -tearoff 0
+  $m add cascade -label $al(MC,filelist) -menu $m.filelist
+  $m.filelist add command -label $al(MC,filelist) -command {alited::bar::BAR popList} -accelerator $al(acc_21)
+  $m.filelist add checkbutton -label [msgcat::mc {Sorted}] -variable alited::al(sortList)
   $m add separator
   $m add command -label $al(MC,colorpicker) -command alited::tool::ColorPicker
   $m add command -label $al(MC,datepicker) -command alited::tool::DatePicker
@@ -307,6 +310,8 @@ proc menu::RunTip {} {
   set m [set al(MENUHELP) $al(WIN).menu.help]
   $m add command -label Tcl/Tk -command alited::tool::Help -accelerator F1
   $m add command -label alited -command alited::HelpAlited
+  $m add separator
+  $m add command -label $al(MC,updateALE) -command {alited::ini::CheckUpdates yes}
   $m add separator
   $m add command -label [msgcat::mc "About..."] -command alited::HelpAbout
   FillRecent